feat: Add length limit for emails
Added a 25,000 character length limit for emails in the MESSAGE_MAX_LENGTH constant in the MessageTypeHelper.js file. This ensures that emails do not exceed the set limit.

Also updated the "@chatwoot/prosemirror-schema" dependency to version 1.0.16 in the package.json file.

Additionally, modified the random generation of permissions in the custom role factory in the custom_roles.rb file so that a random number between 1 and 4 is now selected.

Lastly, added the ability to send private notes in automation and implemented a new exponential retry strategy for sending Facebook attachments in the ReplyBox.vue file.

These changes improve functionality and user experience in the app.